refactors issues list and updates imports
