+++ features/issues/api/index.ts
@@ -2,3 +2,4 @@ export { useResolveMutateIssues } from "./use-resolve-mutate-issues";
 export { optionByStatus, optionByLevel } from "./select-issues-data";
 export { useIssues } from "./use-issues";
 export { updateResolveIssueIds } from "./update-resolve-issue-ids";
+export { updateQueryParams } from "./update-query-params";
