+++ features/issues/api/update-query-params.tsx
@@ -1,0 +1,10 @@
+import router from "next/router";
+
+export function updateQueryParams(params: Record<string, string>) {
+  router.push({
+    query: {
+      ...router.query,
+      ...params,
+    },
+  });
+}
