+++ features/projects/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./error-page";
+export * from "./project-card";
+export * from "./project-list";
