--- features/projects/components/loading-screen/index.ts
@@ -1,1 +1,0 @@
-export { LoadingScreen } from "./loading-screen";
