+++ features/projects/components/project-list/project-list.tsx
@@ -3,7 +3,7 @@ import { ProjectCard } from "../project-card";
 import { useProjects } from "../../api/use-projects";
 import { breakpoint, space } from "@styles/theme";
 import { ErrorPage } from "../error-page";
-import { LoadingScreen } from "../loading-screen";
+import { LoadingSpinner } from "@features/ui";
 
 const List = styled.ul`
   display: grid;
@@ -24,7 +24,7 @@ export function ProjectList() {
   const { data, isLoading, isError, error } = useProjects();
 
   if (isLoading) {
-    return <LoadingScreen />;
+    return <LoadingSpinner />;
   }
 
   if (isError) {
