+++ features/projects/index.ts
@@ -1,3 +1,4 @@
 export * from "./api/use-projects";
 export * from "./components/project-list";
+export * from "./components/error-page";
 export * from "./types/project.types";
