+++ features/ui/index.ts
@@ -8,3 +8,4 @@ export * from "./checkbox";
 export * from "./input";
 export * from "./modal";
 export * from "./select";
+export * from "./loading-spinner";
