+++ features/ui/loading-spinner/index.ts
@@ -1,0 +1,1 @@
+export { LoadingSpinner } from "./loading-spinner";
